// chromesocketxhr.d.ts
import { IOStream } from "./stream";
export declare class ChromeSocketXMLHttpRequest {
    chunks: any;
    responseDataParsed: {
        code: any;
        status: any;
        proto: any;
        headers: {};
    };
    ui8IndexOf(arr: any, s: any, startIndex: any): any;
    onload: any;
    _finished: boolean;
    onerror: any;
    opts: any;
    timedOut: boolean;
    timeout: number;
    timeoutId: any;
    stream: IOStream;
    connecting: boolean;
    writing: boolean;
    haderror: boolean;
    closed: boolean;
    sockInfo: any;
    responseType: any;
    extraHeaders: {};
    headersReceived: boolean;
    responseHeaders: any;
    responseHeadersParsed: any;
    responseBody: any;
    responseLength: any;
    responseBytesRead: any;
    requestBody: any;
    secured: boolean;
    uri: any;
    ontimeout: any;
    constructor();
    open(method: any, url: any, async: any): void;
    setRequestHeader(key: any, val: any): void;
    cancel(): void;
    send(data: any): void;
    createRequestHeaders(): string;
    checkTimeout(): void;
    error(data: any): void;
    onStreamClose(evt: any): void;
    onCreate(sockInfo: any): void;
    onConnect(result: any): void;
    getHost(): any;
    getPort(): number;
    onHeaders(data: any): void;
    onChunkDone(data: any): void;
    getNewChunk(data: any): void;
    onBody(body: any): void;
}